fix(TopicProvider): tests are passing
