+++ app/data_providers/topic_provider.py
@@ -15,7 +15,7 @@ class TopicProvider:
     def __init__(self, aioboto3_session: aioboto3.session.Session):
         self.aioboto3_session = aioboto3_session
 
-    @xray_recorder.capture_async('models_topic_get_all')
+    @xray_recorder.capture_async('TopicProvider.get_all')
     @cached(ttl=60, key='TopicProvider.get_all')
     async def get_all(self) -> List[TopicModel]:
         """
